backend/requirements.txt
fastapi[all]
uvicorn[standard]

sqlalchemy>=2.0
asyncpg

python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

python-dotenv